Implemented compareCandidate()
Minor modifications to NewZealand to add new global to enable the testing suite.
Minor modificaiton to MainController to call the testing suite.
New class added Tester.
Minor change to runtime.html to add tester to the scripts in the header. 
Tester has three methods. constructor, storedCheck which checks if the tester is running in ad-hoc or generated mode, and compareCandidates. None of the modifications have been checked against standardJS