added compareParty()
Added code to compare parties to Tester.js. Party comparison is comprehensive, the program checks all the attributes that are single valued on the party, and then moves on to working through the lists of candidates and mp's, calling compareCandidate() on each item in the list. Comparison values are dynamically generated, and a summary print out to the console occours at the end of the party analysis, dynamically generated around the lists stored in candidateLists.  not JSchecker validated yet.